optional settings_local.py creation

+++ .docker/run.sh
@@ -3,10 +3,11 @@ APP_ROOT=/app
 APP_DIR=$APP_ROOT/snipt
 COOKIE_DOMAIN=${SESSION_COOKIE_DOMAIN:-.snipt.net}
 SECRET_KEY=${SECRET_KEY:-changeme}
-cp $APP_DIR/settings_local-template.py $APP_DIR/settings_local.py
-# replace SESSION_COOKIE_DOMAIN
-sed -i "s/^SESSION_COOKIE_DOMAIN.*/SESSION_COOKIE_DOMAIN = '$COOKIE_DOMAIN'/g" $APP_DIR/settings_local.py
-sed -i "s/^SECRET_KEY.*/SECRET_KEY = '$SECRET_KEY'/g" $APP_DIR/settings_local.py
+if [ ! -e "$APP_DIR/settings_local.py" ]; then
+    cp $APP_DIR/settings_local-template.py $APP_DIR/settings_local.py
+    sed -i "s/^SESSION_COOKIE_DOMAIN.*/SESSION_COOKIE_DOMAIN = '$COOKIE_DOMAIN'/g" $APP_DIR/settings_local.py
+    sed -i "s/^SECRET_KEY.*/SECRET_KEY = '$SECRET_KEY'/g" $APP_DIR/settings_local.py
+fi
 pushd $APP_DIR
 python manage.py syncdb --noinput
 python manage.py migrate --noinput
